Reintroduce household overview as separate table
