--- src/modules/household/components/EditHouseholdMemberTable.tsx
@@ -1,111 +1,0 @@
-import { Box, Tooltip } from '@mui/material';
-import { useMemo } from 'react';
-
-import { HOUSEHOLD_MEMBER_COLUMNS } from './HouseholdMemberTable';
-
-import { CheckIcon } from '@/components/elements/SemanticIcons';
-import GenericTable from '@/components/elements/table/GenericTable';
-import { SsnDobShowContextProvider } from '@/modules/client/providers/ClientSsnDobVisibility';
-import HmisEnum from '@/modules/hmis/components/HmisEnum';
-import { clientBriefName, sortHouseholdMembers } from '@/modules/hmis/hmisUtil';
-import { ManageHouseholdProject } from '@/modules/household/components/ManageHousehold';
-import { useHouseholdMenuActions } from '@/modules/household/hooks/useHouseholdMenuActions';
-import { HmisEnums } from '@/types/gqlEnums';
-import {
-  HouseholdClientFieldsFragment,
-  HouseholdFieldsFragment,
-  RelationshipToHoH,
-} from '@/types/gqlTypes';
-
-interface Props {
-  household: HouseholdFieldsFragment;
-  project: ManageHouseholdProject;
-
-  refetchHousehold: any;
-  loading?: boolean;
-  canEdit?: boolean;
-}
-
-//rename to HouseholdMemberTable
-const EditHouseholdMemberTable = ({
-  household,
-  project,
-  refetchHousehold,
-  loading,
-  canEdit,
-}: Props) => {
-  const currentMembers = useMemo(
-    () => sortHouseholdMembers(household.householdClients),
-    [household]
-  );
-
-  const columns = useMemo(() => {
-    const anyExited = household.householdClients.find(
-      (hc) => !!hc.enrollment.exitDate
-    );
-
-    // TODO ADD: MCI ID
-    return [
-      HOUSEHOLD_MEMBER_COLUMNS.clientName,
-      HOUSEHOLD_MEMBER_COLUMNS.entryDate,
-      ...(anyExited ? [HOUSEHOLD_MEMBER_COLUMNS.exitDate] : []),
-      HOUSEHOLD_MEMBER_COLUMNS.enrollmentStatus,
-      {
-        header: (
-          <Tooltip title='Head of Household' placement='top' arrow>
-            <Box sx={{ width: 'fit-content' }}>HoH</Box>
-          </Tooltip>
-        ),
-        key: 'HoH',
-        render: (hc: HouseholdClientFieldsFragment) =>
-          hc.enrollment.relationshipToHoH ===
-            RelationshipToHoH.SelfHeadOfHousehold && <CheckIcon />,
-      },
-      {
-        header: 'Rel. to HoH',
-        key: 'relationship',
-        render: (hc: HouseholdClientFieldsFragment) => (
-          <HmisEnum
-            value={hc.relationshipToHoH}
-            enumMap={HmisEnums.RelationshipToHoH}
-          />
-        ),
-      },
-      HOUSEHOLD_MEMBER_COLUMNS.dobAge,
-      HOUSEHOLD_MEMBER_COLUMNS.assignedUnit(household.householdClients),
-    ];
-  }, [household.householdClients]);
-
-  const { getRowSecondaryActionConfigs, actionDialogs, actionLoading } =
-    useHouseholdMenuActions({
-      household,
-      refetchHousehold,
-      loading,
-      currentMembers,
-      project,
-    });
-
-  return (
-    <>
-      <SsnDobShowContextProvider>
-        <GenericTable<HouseholdClientFieldsFragment>
-          rows={currentMembers}
-          rowName={(row) => clientBriefName(row.client)}
-          // TODO : only render actions if user can edit household
-          rowSecondaryActionConfigs={
-            canEdit ? getRowSecondaryActionConfigs : undefined
-          }
-          columns={columns}
-          loading={loading || actionLoading}
-          loadingVariant='linear'
-          tableProps={{
-            'aria-label': 'Manage Household',
-          }}
-        />
-      </SsnDobShowContextProvider>
-      {actionDialogs}
-    </>
-  );
-};
-
-export default EditHouseholdMemberTable;
