+++ src/modules/household/components/HouseholdOverviewTable.stories.tsx
@@ -1,20 +1,20 @@
 import { Meta, StoryObj } from '@storybook/react';
 
-import HouseholdMemberTable from './HouseholdMemberTable';
+import HouseholdOverviewTable from './HouseholdOverviewTable';
 
 import { enrollmentWithHoHMock } from '@/test/__mocks__/requests';
 
 export default {
-  component: HouseholdMemberTable,
+  component: HouseholdOverviewTable,
   parameters: {
     dashboardContext: 'enrollment',
     apolloClient: {
       mocks: [enrollmentWithHoHMock],
     },
   },
-} as Meta<typeof HouseholdMemberTable>;
+} as Meta<typeof HouseholdOverviewTable>;
 
-type Story = StoryObj<typeof HouseholdMemberTable>;
+type Story = StoryObj<typeof HouseholdOverviewTable>;
 
 export const Default: Story = {
   args: {
